PLS-134 Feat:Add API for summary
